rename kube pkg to kapi
